LayoutTests/http/tests/xmlviewer/dumpAsText/wml-expected.txt

CONSOLE MESSAGE: line 1: Unsafe attempt to load URL http://www.wapforum.org/DTD/wml_1.1.xml from frame with URL http://127.0.0.1:8000/xmlviewer/dumpAsText/wml.xml. Domains, protocols and ports must match.

This XML file does not appear to have any style information associated with it. The document tree is shown below.
<!--
If you can see this, this test has failed, except if you are not using a WML-aware client.
-->
<?sds erfe?>
<wml>
<header>
This tests that xml viewer is not used for documents with WML doctype.
</header>
<card id="card1" title="Card 1">
<FAIL/>
<p>SUCCESS</p>
</card>
</wml>
